# internal/pkg/service/appsproxy/api/openapi/openapi3.yaml
openapi: 3.0.3
info:
    title: Data application proxy
    description: A service for proxing requests/authorization to data applications using Keboola components.
    version: "1.0"
servers:
    - url: https://appsproxy.{stack}
      variables:
        stack:
            enum:
                - keboola.com
                - eu-central-1.keboola.com
                - north-europe.azure.keboola.com
                - eu-west-1.aws.keboola.dev
                - east-us-2.azure.keboola-testing.com
            default: keboola.com
    - url: http://localhost:8000
paths:
...
    /v1/validate:
        get:
            tags:
                - template
            summary: Validation of OIDC authorization provider configuration
            description: Validation endpoint of OIDC authorization provider configuration.
            operationId: Validate
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Validations'
                            example:
                                configuration:
                                    - exampleerrordata:
                                        statusCode: 404
                                        error: validations incorrect
                                        message: unable to validate due to missing entry
                                      id: test
                                      name: My Template
            security:
                - storage-api-token: []
components:
    schemas:
        Configuration:
            type: object
            properties:
                clientID:
                    type: string
                    description: Client ID of provider.
                    example: github.oidc
                clientSecret:
                    type: string
                    description: Client secret provided by OIDC provider.
                    example: thisissupersecret
                id:
                    type: string
                    description: Unique ID of provider.
                    example: oidc#1
            description: The configuration that is part of the auth providers section.
            example:
                clientID: github.oidc
                clientSecret: thisissupersecret
                id: oidc#1
            required:
                - id
                - clientID
                - clientSecret
...
        Validations:
            type: object
            properties:
                configuration:
                    type: array
                    items:
                        $ref: '#/components/schemas/Configuration'
                    description: All authorization providers.
                    example:
                        - exampleerrordata:
                            statusCode: 404
                            error: validations incorrect
                            message: unable to validate due to missing entry
                          id: test
                          name: My Template
            example:
                configuration:
                    - exampleerrordata:
                        statusCode: 404
                        error: validations incorrect
                        message: unable to validate due to missing entry
                      id: test
                      name: My Template
    securitySchemes:
        storage-api-token:
            type: apiKey
            description: Storage Api Token Authentication.
            name: X-StorageApi-Token
            in: header
...